Fixed distance only using one byte and bug with reading history data

+++ ÄstrandTest/ÄstrandTestFietsClient/BikeCommunication/DataManager.cs
@@ -41,7 +41,7 @@ namespace ÄstrandTestFietsClient.BikeCommunication
             if (this.clientMessage.HasPage16)
                 PushMessage();
 
-            this.clientMessage.Distance = (byte)distance;
+            this.clientMessage.Distance = distance;
             this.clientMessage.Speed = (byte)speed;
             this.clientMessage.HasPage16 = true;
         }
